+++ engines/supernova/supernova.h
@@ -125,12 +125,12 @@ public:
 
 	void add(Object &obj);
 	void remove(Object &obj);
-	Object *get(size_t index) const;
+	Object *get(uint index) const;
 	Object *get(ObjectID id) const;
 
 private:
 	Object *_inventory[kMaxCarry];
-	size_t _numObjects;
+	uint _numObjects;
 };
 
 class GameManager {
